+++ internal/provider/framework_provider.go
@@ -148,5 +148,6 @@ func (p *FrameworkProvider) ListResources(ctx context.Context) []func() list.Lis
 
 	return []func() list.ListResource{
 		newCustomRoleListResource,
+		newUserListResource,
 	}
 }